% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePermutation.R
\name{additionalRunIF}
\alias{additionalRunIF}
\title{Internal Utility Function for Random Sampling Loops}
\usage{
additionalRunIF(x, y, z, p)
}
\value{
No return value. This function is designed to allow large permutation tests distributed to multiple nodes.
}
\description{
Executes nested loops to perform random sampling operations.
This function is intended for internal use only and should not be accessed by package users.
}
\keyword{internal}
