% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMutualInformation.R
\name{getCMI}
\alias{getCMI}
\title{Calculate Conditional Mutual Information}
\usage{
getCMI(x_1, x_2, x_3, bin = NULL, sp_order = NULL)
}
\arguments{
\item{x_1}{A numeric vector for the first variable.}

\item{x_2}{A numeric vector for the second variable. Must match `x_1` length.}

\item{x_3}{A numeric vector for the condition variable. Must match `x_1` length.}

\item{bin}{An integer specifying the number of bins. Default is NULL.}

\item{sp_order}{An integer specifying the spline order. Must be less than `bin`. Default is NULL.}
}
\value{
A numeric value representing the conditional mutual information (CMI).
}
\description{
Computes the conditional mutual information I(x_1; x_2 | x_3)
using the specified number of bins and spline order.
}
\examples{
x_1 <- rnorm(100)
x_2 <- rnorm(100)
x_3 <- rnorm(100)
getCMI(x_1, x_2, x_3)

}
