% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeEntropy.R
\name{getEntropy}
\alias{getEntropy}
\title{Calculate Univariate Entropy}
\usage{
getEntropy(x_1, bin = NULL, sp_order = NULL)
}
\arguments{
\item{x_1}{A numeric vector for the only variable.}

\item{bin}{An integer specifying the number of bins. Default is NULL.}

\item{sp_order}{An integer specifying the spline order. Must be less than `bin`. Default is NULL.}
}
\value{
A numeric value representing the entropy of the vector.
}
\description{
This function calculates the univariate entropy of a numeric vector using the specified
number of bins and spline order.
}
\examples{
x_1 <- rnorm(100)
getEntropy(x_1)

}
