% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeEntropy.R
\name{getEntropyQuadri}
\alias{getEntropyQuadri}
\title{Calculate Joint Entropy for Four Variables}
\usage{
getEntropyQuadri(x_1, x_2, x_3, x_4, bin = NULL, sp_order = NULL)
}
\arguments{
\item{x_1}{A numeric vector for the first variable.}

\item{x_2}{A numeric vector for the second variable. Must be the same length as `x_1`.}

\item{x_3}{A numeric vector for the third variable. Must be the same length as `x_1`.}

\item{x_4}{A numeric vector for the fourth variable. Must be the same length as `x_1`.}

\item{bin}{An integer specifying the number of bins. Default is NULL.}

\item{sp_order}{An integer specifying the spline order. Must be less than `bin`. Default is NULL.}
}
\value{
A numeric value representing the joint entropy of the four vectors.
}
\description{
Computes the joint entropy of four numeric vectors using the specified number of bins and spline order.
}
\examples{
x_1 <- rnorm(100)
x_2 <- rnorm(100)
x_3 <- rnorm(100)
x_4 <- rnorm(100)
getEntropyQuadri(x_1, x_2, x_3, x_4)

}
