% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMatrix.R
\name{MImat2matAll}
\alias{MImat2matAll}
\title{Normalized Mutual Information Between Each Row of Two Matrices}
\usage{
MImat2matAll(mat1, mat2, bin = NULL, sp_order = NULL)
}
\arguments{
\item{mat1}{A numeric matrix. For example, each row represents a gene and each column represents a sample.}

\item{mat2}{Another numeric matrix to compare against. Must have the columns as `mat1`.}

\item{bin}{An integer specifying the number of bins. Default is NULL.}

\item{sp_order}{An integer specifying the spline order. Must be less than `bin`. Default is NULL.}
}
\value{
A matrix where each element corresponds to the normalized mutual information (MI) between
respective rows of `mat1` and `mat2`.
}
\description{
Computes the normalized mutual information (MI) between each row of first matrix and each row of second matrix,
normalized by their individual information content, using the specified number of bins and spline order.
}
\examples{
mat1 <- matrix(rnorm(1000), nrow = 10, ncol = 100)
mat2 <- matrix(rnorm(2000), nrow = 200, ncol = 100)
MImat2matAll(mat1, mat2)

}
