% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePermutation.R
\name{permute_within_bins}
\alias{permute_within_bins}
\title{Internal Utility Function for Within-Bin Permutation}
\usage{
permute_within_bins(x, y, bin = 8)
}
\arguments{
\item{x}{A numeric vector used to define bins.}

\item{y}{A vector with the same length as `x`. Values in `y` are permuted within bins of `x`.}

\item{bin}{An integer specifying the number of evenly spaced bins. Default is 8.}
}
\value{
A vector with the same length as `y`, where values have been permuted within bins of `x`.
}
\description{
Permutes the values of one vector within bins defined by another vector.
This function is intended for internal use only and should not be accessed by package users.

The binning vector is split into evenly spaced bins between its minimum and maximum
values. Values of the target vector are then randomly permuted only within each bin.
This preserves the approximate relationship between the target vector and the binning
variable while breaking the original sample-wise ordering within local strata.
}
\keyword{internal}
