% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_params.R
\name{resolve_params}
\alias{resolve_params}
\title{Internal Utility Function for Parameter Selection}
\usage{
resolve_params(n, bin = NULL, sp_order = NULL)
}
\arguments{
\item{n}{Integer. Length of the input vector.}

\item{bin}{Number or \code{NULL}. Number of bins.}

\item{sp_order}{Number or \code{NULL}. Spline order.}
}
\value{
A list containing the validated or automatically generated values
\code{bin} and \code{sp_order}.
}
\description{
Resolves and validates the \code{bin} and \code{sp_order} parameters.
If \code{bin} is not provided or valid, both \code{bin} and \code{sp_order} are generated automatically.
If \code{bin} is valid but \code{sp_order} is not provided or valid, only \code{sp_order} is generated.
This function is intended for internal use only and should not be accessed by package users.
}
\keyword{internal}
